requirements.txt
fastapi==0.111.0 # YTEST: FastAPI 버전 고정
uvicorn==0.30.1  # YTEST: ASGI 서버
pydantic==2.7.4  # YTEST: 데이터 유효성 검사 및 설정 관리
pydantic-settings==2.2.1 # YTEST: .env 파일에서 설정 로드
mysql-connector-python==8.4.0 # YTEST: MySQL 연결 드라이버
